feat: restructure environment handling by introducing EnvModule and EnvService
