+++ src/infra/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { AuthModule } from './auth/auth.module'
-import { envSchema } from './env'
+import { envSchema } from './env/env'
 import { HttpModule } from './http/http.module'
-import { EnvService } from './env.service'
+import { EnvModule } from './env/env.module'
 
 @Module({
   imports: [
@@ -13,7 +13,7 @@ import { EnvService } from './env.service'
     }),
     AuthModule,
     HttpModule,
+    EnvModule,
   ],
-  providers: [EnvService],
 })
 export class AppModule {}
